+++ src/auth/guards/access-token.guard.ts
@@ -1,11 +1,5 @@
-import * as process from 'process';
 import { Reflector } from '@nestjs/core';
-import {
-  CanActivate,
-  ExecutionContext,
-  Injectable,
-  UnauthorizedException,
-} from '@nestjs/common';
+import { CanActivate, ExecutionContext, Injectable } from '@nestjs/common';
 import { JwtService } from '@nestjs/jwt';
 import { AuthGuard } from '@nestjs/passport';
 import { IS_PUBLIC_KEY } from '../../app.decorators';
@@ -20,7 +14,7 @@ export class AccessTokenGuard
     super();
   }
 
-  async canActivate(context: ExecutionContext): Promise<boolean> {
+  canActivate(context: ExecutionContext) {
     const isPublic = this.reflector.getAllAndOverride<boolean>(IS_PUBLIC_KEY, [
       context.getHandler(),
       context.getClass(),
@@ -30,27 +24,6 @@ export class AccessTokenGuard
       return true;
     }
 
-    const request = context.switchToHttp().getRequest();
-    const accessToken = this.extractTokenFromHeader(request);
-
-    if (!accessToken) {
-      throw new UnauthorizedException();
-    }
-
-    try {
-      request['user'] = await this.jwtService.verifyAsync(accessToken, {
-        secret: process.env.JWT_SECRET_KEY,
-      });
-    } catch (e) {
-      throw new UnauthorizedException();
-    }
-
-    return true;
-  }
-
-  private extractTokenFromHeader(request: Request): string | undefined {
-    const [type, token] = request.headers['authorization']?.split(' ') ?? [];
-
-    return type === 'Bearer' ? token : undefined;
+    return super.canActivate(context);
   }
 }
